+++ tea/runtimeDataStructures/dataset.py
@@ -7,6 +7,12 @@ from pathlib import Path
 from urllib.parse import urlparse
 import requests
 
+from .variables import (ordinal, isordinal,
+                    nominal, isnominal,
+                    ratio, isratio,
+                    interval, isinterval
+                   )
+
 BASE_PATH = os.getcwd()
 
 def _dir_exists(path):
@@ -14,14 +20,14 @@ def _dir_exists(path):
 
 @attr.s(hash=True)
 class Dataset(object): 
-    dfile = attr.ib() # path name 
-    variables = attr.ib() # list of Variable objects <-- TODO: may not need this in new implementation....
-    pid_col_name = attr.ib() # name of column in pandas DataFrame that has participant ids
+    dfile = attr.ib(default=None,kw_only=True) # path name 
+    variables = attr.ib(init=False) # list of Variable objects <-- TODO: may not need this in new implementation....
+    pid_col_name = attr.ib(default=None,kw_only=True) # name of column in pandas DataFrame that has participant ids
     row_pids = attr.ib(init=False) # list of unique participant ids
     data = attr.ib(init=False) # pandas DataFrame
     
     @staticmethod
-    def load(path: str, name):
+    def load_url(path: str, name):
         assert(isinstance(path, str))
         home = Path.home()
         tea_path = home / '.tea'
@@ -56,7 +62,8 @@ class Dataset(object):
 
         return csv_path
 
-        
+    def load_df(self, df: pd.DataFrame):
+        self.data = df
 
     def __attrs_post_init__(self): 
         if self.dfile: 
@@ -73,6 +80,35 @@ class Dataset(object):
         for v in self.variables: # checks that the Variable is known to the Dataset object
             if v.name == var_name: 
                 return self.data[var_name] # returns the data, not the variable object
+
+    def set_variables(self, vars: Dict[str, str]):
+        var_dtype = 'data type'
+        var_name = 'name'
+        var_dtype = 'data type'
+        var_categories = 'categories'
+        var_drange = 'range'
+        self.variables = []
+
+        for var in vars: 
+            name = var['name']
+
+            if (var[var_dtype] == 'nominal'): 
+                categories = var[var_categories]
+                v_obj = nominal(name, categories)
+            elif (var[var_dtype] == 'ordinal'): 
+                categories = var[var_categories]
+                v_obj = ordinal(name, categories)
+            elif (var[var_dtype] == 'interval'): 
+                drange=None
+                if var_drange in var: 
+                    drange = var[var_drange]
+                v_obj = interval(name, drange)
+            else: 
+                assert(var[var_dtype] == 'ratio')
+                drange = var[var_drange] if var_drange in var else None
+                v_obj = interval(name, drange)
+
+            self.variables.append(v_obj)
 
     # Returns variable object -- may want to altogether replace the __getitem__
     def get_variable(self, var_name: str):
